+++ libs/@local/hashql/core/Cargo.toml
@@ -45,4 +45,5 @@ workspace = true
 
 [dev-dependencies]
 insta     = { workspace = true }
+rstest    = "0.25.0"
 test-case = "3.3.1"
