+++ libs/@local/hashql/core/src/type/collection/type_id_set.rs
@@ -167,7 +167,7 @@ mod test {
             collection::TypeIdSet,
             environment::Environment,
             kind::{TypeKind, primitive::PrimitiveType, test::primitive},
-            test::instantiate,
+            tests::instantiate,
         },
     };
