+++ libs/@local/hashql/core/src/type/kind/intrinsic.rs
@@ -874,7 +874,7 @@ mod tests {
                 union::UnionType,
             },
             lattice::{Lattice as _, test::assert_lattice_laws},
-            test::{instantiate, instantiate_infer, instantiate_param},
+            tests::{instantiate, instantiate_infer, instantiate_param},
         },
     };
