+++ libs/@local/hashql/core/src/type/kind/struct.rs
@@ -608,7 +608,7 @@ mod test {
                 union::UnionType,
             },
             lattice::{Lattice as _, Projection, Subscript, test::assert_lattice_laws},
-            test::{instantiate, instantiate_infer, instantiate_param},
+            tests::{instantiate, instantiate_infer, instantiate_param},
         },
     };
