+++ libs/@local/hashql/core/src/type/kind/tuple.rs
@@ -425,7 +425,7 @@ mod test {
                 union::UnionType,
             },
             lattice::{Lattice as _, Projection, Subscript, test::assert_lattice_laws},
-            test::{instantiate, instantiate_infer, instantiate_param},
+            tests::{instantiate, instantiate_infer, instantiate_param},
         },
     };
