+++ libs/@local/hashql/core/src/type/kind/union.rs
@@ -770,7 +770,7 @@ mod test {
                 tuple::TupleType,
             },
             lattice::{Lattice as _, Projection, Subscript, test::assert_lattice_laws},
-            test::{instantiate, instantiate_infer, instantiate_param},
+            tests::{instantiate, instantiate_infer, instantiate_param},
         },
     };
